+++ src/action/Action.js
@@ -1,12 +1,33 @@
 import { ACTION_BASE } from './ACTION_TYPE';
 
-export default class Action {
+/**
+ * Action to interact with
+ *
+ * @property {string} name Name
+ * @property {string} description Description
+ * @property {ACTION_TYPE} type Type
+ */
+class Action {
+  /**
+   * Creates an instance of Action.
+   * @param {string} [name=''] Name
+   * @param {string} [description=''] Description
+   *
+   * @memberOf Action
+   */
   constructor(name = '', description = '') {
     this.name = name;
     this.description = description;
     this.type = ACTION_BASE;
   }
 
+  /**
+   * Test if action is valid
+   *
+   * @returns {boolean}
+   *
+   * @memberOf Action
+   */
   isValid() {
     if (this.name) {
       return true;
@@ -14,3 +35,5 @@ export default class Action {
     return false;
   }
 }
+
+export default Action;